# routes/swagger.yaml
openapi: 3.0.0
info:
  title: PocketMoney API
  version: 1.0.0
  description: API for PocketMoney application
servers:
  - url: 'http://localhost:3001'
    description: 'localhost'
  - url: 'http://34.170.36.196/api/'
    description: 'Google cloud'
paths:
...
  /budget/update/{id}:
    put:
      summary: Update existing budget
      description:
      tags:
        - budget
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            minimum: 1
          description: ID of the budget to be updated
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  description: Name of the budget to be updated
      responses:
        200:
          description: Success
  /budget/delete/{id}:
    delete:
      summary: Delete budget
      description:
      tags:
        - budget
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            minimum: 1
          description: ID of the budget to be deleted
      responses:
        200:
          description: Success
...
  /budget/expense/add/{budgetid}:
    put:
      summary: Add budget expense
      tags:
        - budget/expense
      parameters:
        - in: path
          name: budgetid
          required: true
          schema:
            type: string
            minimum: 1
          description: ID of the budget into which the expense is being added
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                expenseAccount:
                  type: string
                  description: Account from which the expense will be paid
                  enum: [Unicredit, Revolut, Kreditka, Hotovost]
                expenseName:
                  type: string
                  description: Name of the expense
                expenseType:
                  type: string
                  description: Type of the expense
                  enum: [Tax, Mandatory, Must, Others]
                paidAmount:
                  type: integer
                  description: Amount of the expense which was already paid
                plannedAmount:
                  type: integer
                  description: Amount of the expense planned at the beginning of the budget period
                remainingAmount:
                  type: integer
                  description: Remaining amount of the expense. It is changing as the expense is being paid during the budget period.
      responses:
        200:
          description: Success
  /budget/expense/update/{budgetid}:
    put:
      summary: Update budget expense
      tags:
        - budget/expense
      parameters:
        - in: path
          name: budgetid
          required: true
          schema:
            type: string
            minimum: 1
          description: ID of the budget in which the expense is being updated
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                expenseId:
                  type: string
                  description: ID of the expense being updated
                expenseAccount:
                  type: string
                  description: Account from which the expense will be paid
                  enum: [Unicredit, Revolut, Kreditka, Hotovost]
                expenseName:
                  type: string
                  description: Name of the expense
                expenseType:
                  type: string
                  description: Type of the expense
                  enum: [Tax, Mandatory, Must, Others]
                paidAmount:
                  type: integer
                  description: Amount of the expense which was already paid
                plannedAmount:
                  type: integer
                  description: Amount of the expense planned at the beginning of the budget period
                remainingAmount:
                  type: integer
                  description: Remaining amount of the expense. It is changing as the expense is being paid during the budget period.
      responses:
        200:
          description: Success
  /budget/expense/delete/{budgetid}:
    put:
      summary: Delete budget expense
      tags:
        - budget/expense
      parameters:
        - in: path
          name: budgetid
          required: true
          schema:
            type: string
            minimum: 1
          description: ID of the budget in which the expense is being deleted
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                expenseId:
                  type: string
                  description: ID of the expense being deleted
      responses:
        200:
          description: Success
...
  /config/mandatoryExpenses/new:
    post:
      summary: Add new Mandatory expense
      tags:
        - config/mandatoryExpenses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                expenseName:
                  type: string
                expenseAmount:
                  type: string
                expenseAccount:
                  type: string
                  enum: [Unicredit, Revolut, Kreditka, Hotovost]
      responses:
        200:
          description: Success
  /config/mandatoryExpenses/update/{id}:
    put:
      summary: Update Mandatory expense
      description:
      tags:
        - config/mandatoryExpenses
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            minimum: 1
          description: ID of the Mandatory expense being updated
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#components/schemas/MandatoryExpense'
      responses:
        200:
          description: Success
  /config/mandatoryExpenses/delete/{id}:
    delete:
      summary: Delete Mandatory expense
      description:
      tags:
        - config/mandatoryExpenses
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            minimum: 1
          description: ID of the Mandatory expense being deleted
      responses:
        200:
          description: Success
...
  /config/monthlyExpenses/new:
    post:
      summary: Creates new planned repeating monthly expense
      tags:
        - config/monthlyExpenses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#components/schemas/MonthlyExpense'
      responses:
        200:
          description: Success
  /config/monthlyExpenses/update/{id}:
    put:
      summary: Updates repeating monthly expense
      tags:
        - config/monthlyExpenses
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            minimum: 1
          description: ID of the Monthly expense being updated
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#components/schemas/MonthlyExpense'
      responses:
        200:
          description: Success
        400:
          description: Wrong ID in path
  /config/monthlyExpenses/delete/{id}:
    delete:
      summary: Deletes Monthly repeating expense
      tags:
        - config/monthlyExpenses
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            minimum: 1
          description: ID of the Monthly expense being deleted
      responses:
        200:
          description: Success
        400:
          description: Wrong ID in path
        500:
          description: Error when deleting expense
...